refactor(analytics): move generator to internal
