+++ go/analytics/internal/generator/fraud.go
