+++ go/analytics/internal/generator/fraud_test.go
