+++ go/analytics/internal/generator/generator.go
