+++ go/analytics/internal/generator/generator_test.go
