+++ go/analytics/internal/generator/rng.go
