+++ go/analytics/internal/generator/rng_test.go
